Changing taxes.cs types to float

+++ Energy Saver/Model/Taxes.cs
@@ -6,14 +6,19 @@ namespace Energy_Saver.Model
     {
         [JsonProperty("id")]
         public int Id { get; set; }
+
         [JsonProperty("month")]
         public string Month { get; set; }
+
         [JsonProperty("gasPrice")]
         public float GasPrice { get; set; }
+
         [JsonProperty("electricityPrice")]
         public float ElectricityPrice { get; set; }
+
         [JsonProperty("waterPrice")]
-        public int WaterPrice { get; set; }
+        public float WaterPrice { get; set; }
+
         [JsonProperty("heatingPrice")]
         public float HeatingPrice { get; set; }
